pyproject.toml
[project]
name = "pocket-attorney-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.2",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.26",
    "numpy>=2.4.4",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "uvicorn>=0.42.0",
    "requests",
    "beautifulsoup4",
    "python-multipart",
    "typing-extensions>=4.15.0",
    "pypdf>=6.10.2",
    "redis>=7.4.0",
    "rank-bm25>=0.2.2",
    "sqlalchemy>=2.0.49",
    "psycopg2-binary>=2.9.12",
    "passlib[bcrypt]>=1.7.4",
    "python-jose>=3.5.0",
    "email-validator>=2.3.0",
    "bcrypt==4.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]

[project.scripts]
build-index = "app.ingestion.build_index"

[dependency-groups]
dev = [
    "ruff>=0.15.12",
]
